split Info.jsx
