+++ Employee-Client/src/pages/info/utls.js
@@ -1,0 +1,1 @@
+export const INPUT_SX = { mb: "8px", mr: "4px" };
